+++ lithops/standalone/keeper.py
@@ -15,13 +15,12 @@
 # limitations under the License.
 #
 
-import json
 import os
 import time
 import threading
 import logging
 from lithops.standalone import StandaloneHandler
-from lithops.constants import SA_DATA_FILE, JOBS_DIR
+from lithops.constants import JOBS_DIR
 from lithops.standalone.utils import JobStatus
 
 
@@ -32,7 +31,7 @@ class BudgetKeeper(threading.Thread):
     """
     BudgetKeeper class used to automatically stop the VM instance
     """
-    def __init__(self, config, stop_callback=None, delete_callback=None):
+    def __init__(self, config, instance_data, stop_callback=None, delete_callback=None):
         threading.Thread.__init__(self)
         self.last_usage_time = time.time()
 
@@ -46,9 +45,6 @@ class BudgetKeeper(threading.Thread):
 
         self.runing = False
         self.jobs = {}
-
-        with open(SA_DATA_FILE, 'r') as ad:
-            instance_data = json.load(ad)
 
         self.standalone_handler = StandaloneHandler(self.standalone_config)
         self.instance = self.standalone_handler.backend.get_instance(**instance_data)